Documentação da API

+++ src/main/java/com/equipeturma862/cadastronf/repository/TipoNaturezaOperacaoRepository.java
@@ -3,5 +3,5 @@ package com.equipeturma862.cadastronf.repository;
 import com.equipeturma862.cadastronf.domain.TipoNaturezaOperacao;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TipoNaturezaOperacaoRepository extends CrudRepository<TipoNaturezaOperacao, Integer> {
+public interface TipoNaturezaOperacaoRepository extends CrudRepository<TipoNaturezaOperacao, Long> {
 }
